% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brimr.R
\name{brimr_dept_data}
\alias{brimr_dept_data}
\title{BRIMR Department Data}
\usage{
brimr_dept_data(year, nih_dept_combining_name)
}
\arguments{
\item{year}{desired year}

\item{nih_dept_combining_name}{desired department}
}
\value{
a tibble containing summary information regarding NIH funding
}
\description{
BRIMR Department Data
}
\examples{
brimr_dept_data(2021, "PEDIATRICS")
}
