% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brimr.R
\name{brimr_dept_flextable}
\alias{brimr_dept_flextable}
\title{BRIMR Department Flextable}
\usage{
brimr_dept_flextable(data)
}
\arguments{
\item{data}{tibble containing years in rows and four variables at a minimum:
year, wusm_rank, wusm_n_awards, wusm_prop_nih_total_funding}
}
\value{
a flextable listing rank, number of grants, and percent of NIH
funding for each year
}
\description{
BRIMR Department Flextable
}
