% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brimr.R
\name{brimr_dept_plot}
\alias{brimr_dept_plot}
\title{BRIMR Department Plot}
\usage{
brimr_dept_plot(
  year,
  nih_dept_combining_name,
  y2_max = 0.1,
  y2_by = 0.02,
  nih_pct_lab_nudge_y = 0
)
}
\arguments{
\item{year}{desired year}

\item{nih_dept_combining_name}{desired department}

\item{y2_max}{maximum proportion for second  axis}

\item{y2_by}{increment of second axis labels}

\item{nih_pct_lab_nudge_y}{vertical adjustment to nudge NIH percent labels by}
}
\value{
a ggplot
}
\description{
BRIMR Department Plot
}
\examples{
brimr_dept_plot(2021, "PEDIATRICS")
}
