% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brimr.R
\name{brimr_dept_slide}
\alias{brimr_dept_slide}
\title{BRIMR Department Slide}
\usage{
brimr_dept_slide(
  year,
  nih_dept_combining_name,
  target = brimr_dept_slide_name(nih_dept_combining_name, year)
)
}
\arguments{
\item{year}{desired year}

\item{nih_dept_combining_name}{desired department}

\item{target}{path to the pptx file to write}
}
\description{
BRIMR Department Slide
}
