% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brimr.R
\name{brimr_dept_slide_name}
\alias{brimr_dept_slide_name}
\title{BRIMR Slide Name}
\usage{
brimr_dept_slide_name(nih_dept_combining_name, year)
}
\arguments{
\item{nih_dept_combining_name}{department}

\item{year}{year}
}
\value{
a standardized slide name
}
\description{
BRIMR Slide Name
}
\examples{
brimr_dept_slide_name("PEDIATRICS", 2021)
}
