% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brimr.R
\name{brimr_download_table_3}
\alias{brimr_download_table_3}
\title{BRIMR Download Table 3}
\usage{
brimr_download_table_3(
  year,
  dest = sprintf("MedicalSchoolsOnly_\%s.xls", year)
)
}
\arguments{
\item{year}{desired year}

\item{dest}{destination file}
}
\value{
destination file path
}
\description{
The \href{https://www.brimr.org/}{Blue Ridge Institute for Medical Research}
compiles and disseminates NIH funding data annually. This function attempts
to download Table 3 Medical Schools Only File 10 Mb; the comprehensive file
for all US Medical Schools receiving NIH funds for the desired year.
}
