% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brimr.R
\name{brimr_genetics_data}
\alias{brimr_genetics_data}
\title{BRIMR Genetics Data}
\usage{
brimr_genetics_data(year)
}
\arguments{
\item{year}{desired year}
}
\value{
a 2 x 6 tibble with rows for dept_name either GENETICS or
GENOME INSTITUTE and containing columsn for year, organization_name,
nih_dept_combining_name, dept_name, fisc_year, and wusm_funding.
}
\description{
BRIMR Genetics Data
}
\examples{
brimr_genetics_data(2021)
}
