% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brimr.R
\name{brimr_genetics_flextable}
\alias{brimr_genetics_flextable}
\title{BRIMR Genetics Flextable}
\usage{
brimr_genetics_flextable(data)
}
\arguments{
\item{data}{tibble containing years in rows and three variables at a minimum:
fisc_year, dept_name, wusm_funding. dept_name should contain values of either
GENETICS or GENOME INSTITUTE.}
}
\value{
a flextable listing funding totals for each dept_name for each year
}
\description{
BRIMR Genetics Flextable
}
