% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brimr.R
\name{brimr_genetics_plot}
\alias{brimr_genetics_plot}
\title{BRIMR Genetics Plot}
\usage{
brimr_genetics_plot(year, y2_max = 0.25, y2_by = 0.05, nih_pct_lab_nudge_y = 0)
}
\arguments{
\item{year}{desired year}

\item{y2_max}{maximum proportion for second  axis}

\item{y2_by}{increment of second axis labels}

\item{nih_pct_lab_nudge_y}{vertical adjustment to nudge NIH percent labels by}
}
\value{
a ggplot
}
\description{
BRIMR Genetics Plot
}
\examples{
brimr_genetics_plot(2021)
}
