% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brimr.R
\name{brimr_genetics_slide}
\alias{brimr_genetics_slide}
\title{BRIMR Genetics Slide}
\usage{
brimr_genetics_slide(year, target = brimr_dept_slide_name("GENETICS", year))
}
\arguments{
\item{year}{desired year}

\item{target}{path to the pptx file to write}
}
\description{
BRIMR Genetics Slide
}
