% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brimr.R
\name{brimr_get_fig_data}
\alias{brimr_get_fig_data}
\title{BRIMR Get Fig Data}
\usage{
brimr_get_fig_data(year, nih_dept_combining_name)
}
\arguments{
\item{year}{desired year}

\item{nih_dept_combining_name}{desired department}
}
\value{
a tibble containing summary information regarding NIH funding
}
\description{
BRIMR Get Fig Data
}
\examples{
brimr_get_fig_data(2021, "PEDIATRICS")
}
