% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brimr.R
\name{brimr_make_slide}
\alias{brimr_make_slide}
\title{BRIMR Make PowerPoint Slide}
\usage{
brimr_make_slide(
  year,
  nih_dept_combining_name,
  target = brimr_slide_name(nih_dept_combining_name, year)
)
}
\arguments{
\item{year}{desired year}

\item{nih_dept_combining_name}{desired department}

\item{target}{destination file}
}
\description{
BRIMR Make PowerPoint Slide
}
