% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brimr.R
\name{brimr_map_wusm_dept}
\alias{brimr_map_wusm_dept}
\title{BRIMR Map WUSM Department Name}
\usage{
brimr_map_wusm_dept(nih_dept_combining_name, sanitize = FALSE)
}
\arguments{
\item{nih_dept_combining_name}{standardized NIH department name}

\item{sanitize}{replace spaces and ampersands}
}
\value{
the corresponding WUSM department name
}
\description{
BRIMR Map WUSM Department Name
}
\examples{
brimr_map_wusm_dept("EMERGENCY MEDICINE")
}
