% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brimr.R
\name{brimr_ranking_flextable}
\alias{brimr_ranking_flextable}
\title{BRIMR Ranking Flextable}
\usage{
brimr_ranking_flextable(year)
}
\arguments{
\item{year}{desired year to be compared to the previous}
}
\value{
a printable flextable version of brimr_ranking_table(year)
}
\description{
BRIMR Ranking Flextable
}
\examples{
brimr_ranking_flextable(2021)
}
