% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brimr.R
\name{brimr_ranking_slide}
\alias{brimr_ranking_slide}
\title{BRIMR Ranking Slide}
\usage{
brimr_ranking_slide(
  year,
  target = sprintf("BRIMR_Department_Rankings_\%s.pptx", year)
)
}
\arguments{
\item{year}{desired year to be compared to the previous}

\item{target}{path to the pptx file to write}
}
\description{
BRIMR Ranking Slide
}
