% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brimr.R
\name{brimr_ranking_table}
\alias{brimr_ranking_table}
\title{BRIMR Ranking Table}
\usage{
brimr_ranking_table(year)
}
\arguments{
\item{year}{desired year to be compared to the previous}
}
\value{
a tibble containing data that compares WUSM departments ranks and
funding vs the top funded department for the provided and previous year.
}
\description{
BRIMR Ranking Table
}
\examples{
brimr_ranking_table(2021)
}
