% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brimr.R
\name{brimr_ranking_table_row}
\alias{brimr_ranking_table_row}
\title{BRIMR Ranking Table Row}
\usage{
brimr_ranking_table_row(year, nih_dept_combining_name)
}
\arguments{
\item{year}{desired year}

\item{nih_dept_combining_name}{the NIH combining name of the department}
}
\value{
a tibble containing data for one row in a ranking table
}
\description{
BRIMR Ranking Table Row
}
\examples{
josp:::brimr_ranking_table_row(2021, "PEDIATRICS")
}
