% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brimr_table_3}
\alias{brimr_table_3}
\title{BRIMR Table 3}
\format{
A tibble with compiled Table 3 data from BRIMR.
}
\usage{
brimr_table_3
}
\description{
Compiled NIH Funding data by the Blue Ridge Institute for Medical Research
from their Table 3 dated 2011 to 2021.
}
\keyword{datasets}
