% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brimr_wusm_dept_mappings}
\alias{brimr_wusm_dept_mappings}
\title{BRIMR WUSM Department Mappings}
\format{
A tibble with two columns \code{wusm_dept} which are mapped to
\code{nih_combining_name}.
}
\usage{
brimr_wusm_dept_mappings
}
\description{
A map from WUSM departments to the NIH combining names.
}
\keyword{datasets}
