% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{cagr}
\alias{cagr}
\title{Compound Annual Growth Rate}
\usage{
cagr(begin, final, years)
}
\arguments{
\item{begin}{starting value}

\item{final}{ending value}

\item{years}{number of years}
}
\value{
compound annual growth rate
}
\description{
Compound Annual Growth Rate
}
\examples{
cagr(9000, 13000, 3)
}
