% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{resource}
\alias{resource}
\title{Generate Resource Paths}
\usage{
resource(path, file)
}
\arguments{
\item{path}{relative component of path between josp libary and file}

\item{file}{name of resource}
}
\value{
absolute path to resource file
}
\description{
Generate Resource Paths
}
\examples{
resource("officer", "template_16x9.pptx")
}
