% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brimr.R
\name{brimr_make_slide}
\alias{brimr_make_slide}
\title{BRIMR Make PowerPoint Slide}
\usage{
brimr_make_slide(
  year,
  dept,
  target = sprintf("brimr_\%s_\%s.pptx", dept, year)
)
}
\arguments{
\item{year}{desired year}

\item{dept}{desired department}

\item{target}{destination file}
}
\description{
BRIMR Make PowerPoint Slide
}
